feat(ansible): support for ansible-vault encrypted Kube Secret files
