+++ lib/types.go
@@ -9,3 +9,9 @@ type Result struct {
 	JSON  map[string]interface{}
 	Error error
 }
+
+type KubeSecret struct {
+	Kind string
+	Type string
+	Data map[string]interface{}
+}
