+++ main.go
@@ -21,6 +21,9 @@ func main() {
 				Name:  "Derek Smith",
 				Email: "dsmith@goodwaygroup.com",
 			},
+			&cli.Author{
+				Name: info.AppRepoOwner,
+			},
 		},
 		Copyright:            "(c) 2020 Goodway Group",
 		HelpName:             info.AppName,
@@ -37,13 +40,13 @@ func main() {
 				},
 			},
 			{
-				Name:      "view",
-				Usage:     "View Env local or within a namespace",
-				UsageText: info.ViewCommandHelpText,
+				Name:    "local",
+				Aliases: []string{"l"},
+				Usage:   "Interact with local env files",
 				Subcommands: []*cli.Command{
 					{
-						Name:      "local",
-						Aliases:   []string{"l"},
+						Name:      "view",
+						Aliases:   []string{"v"},
 						Usage:     "View values based on local settings",
 						UsageText: info.ViewLocalCommandHelpText,
 						Flags: []cli.Flag{
@@ -64,8 +67,41 @@ func main() {
 						Action: cmd.ViewLocalEnv,
 					},
 					{
-						Name:      "namespace",
-						Aliases:   []string{"ns"},
+						Name:      "ansible",
+						Aliases:   []string{"legacy", "a"},
+						Usage:     "View value from ansible-vault encrypted Kube Secret file.",
+						UsageText: info.ViewAnsibleEncryptedEnvHelpText,
+						Flags: []cli.Flag{
+							&cli.StringFlag{
+								Name:     "vault-password-file",
+								Usage:    "vault password file `VAULT_PASSWORD_FILE`",
+								Required: false,
+							},
+							&cli.StringFlag{
+								Name:     "encrypted-env-file",
+								Aliases:  []string{"e"},
+								Usage:    "Path to encrypted Kube Secret file",
+								Required: true,
+							},
+							&cli.StringFlag{
+								Name:    "accessor",
+								Aliases: []string{"a"},
+								Usage:   "Accessor key to pull data out of Data block.",
+								Value:   ".env",
+							},
+						},
+						Action: cmd.ViewAnsibleEncryptedEnv,
+					},
+				},
+			},
+			{
+				Name:    "namespace",
+				Aliases: []string{"ns"},
+				Usage:   "Interact with env on a running Pod within a Namespace",
+				Subcommands: []*cli.Command{
+					{
+						Name:      "view",
+						Aliases:   []string{"v"},
 						Usage:     "View values configured withing a namespace",
 						UsageText: info.ViewNamespaceCommandHelpText,
 						Flags: []cli.Flag{
@@ -97,54 +133,92 @@ func main() {
 						},
 						Action: cmd.ViewNamespaceEnv,
 					},
-					{
-						Name:      "diff",
-						Aliases:   []string{"d"},
-						Usage:     "View diff of local vs. namespace",
-						UsageText: info.ViewDiffCommandHelpText,
-						Flags: []cli.Flag{
-							&cli.StringFlag{
-								Name:     "secrets",
-								Aliases:  []string{"s"},
-								Usage:    "Path to secrets.yml",
-								Required: false,
-								Value:    ".docker/secrets.yml",
-							},
-							&cli.StringFlag{
-								Name:     "configmap",
-								Aliases:  []string{"c"},
-								Usage:    "Path to configmap.yaml",
-								Required: true,
-							},
-							&cli.StringFlag{
-								Name:     "namespace",
-								Aliases:  []string{"n"},
-								Usage:    "Kube Namespace list Pods from",
-								Required: true,
-							},
-							&cli.StringFlag{
-								Name:     "cmd",
-								Usage:    "Command to inspect",
-								Required: false,
-								Value:    "node",
-							},
-							&cli.StringFlag{
-								Name:     "filter-prefix",
-								Aliases:  []string{"f"},
-								Usage:    "List of prefixes (csv) used to filter values from display. Set to \"\" to remove any filters.",
-								Required: false,
-								Value:    "npm_,KUBERNETES_,API_PORT",
-							},
-							&cli.StringFlag{
-								Name:     "exclude",
-								Usage:    "List (csv) of specific env vars to exclude values from display. Set to \"\" to remove any exclusions.",
-								Required: false,
-								Value:    "PATH,SHLVL,HOSTNAME",
-							},
-						},
-						Action: cmd.ViewEnvDiff,
+				},
+			},
+			{
+				Name:      "diff",
+				Aliases:   []string{"d"},
+				Usage:     "View diff of local vs. namespace",
+				UsageText: info.ViewDiffCommandHelpText,
+				Flags: []cli.Flag{
+					&cli.StringFlag{
+						Name:     "secrets",
+						Aliases:  []string{"s"},
+						Usage:    "Path to secrets.yml",
+						Required: false,
+						Value:    ".docker/secrets.yml",
+					},
+					&cli.StringFlag{
+						Name:     "configmap",
+						Aliases:  []string{"c"},
+						Usage:    "Path to configmap.yaml",
+						Required: true,
+					},
+					&cli.StringFlag{
+						Name:     "namespace",
+						Aliases:  []string{"n"},
+						Usage:    "Kube Namespace list Pods from",
+						Required: true,
+					},
+					&cli.StringFlag{
+						Name:     "cmd",
+						Usage:    "Command to inspect",
+						Required: false,
+						Value:    "node",
+					},
+					&cli.StringFlag{
+						Name:     "filter-prefix",
+						Aliases:  []string{"f"},
+						Usage:    "List of prefixes (csv) used to filter values from display. Set to \"\" to remove any filters.",
+						Required: false,
+						Value:    "npm_,KUBERNETES_,API_PORT",
+					},
+					&cli.StringFlag{
+						Name:     "exclude",
+						Usage:    "List (csv) of specific env vars to exclude values from display. Set to \"\" to remove any exclusions.",
+						Required: false,
+						Value:    "PATH,SHLVL,HOSTNAME",
 					},
 				},
+				Action: cmd.ViewEnvDiff,
+			},
+			{
+				Name:      "diff:legacy",
+				Aliases:   []string{"diff:ansible"},
+				Usage:     "View diff of local (ansible encrypted) vs. namespace",
+				UsageText: info.ViewDiffCommandHelpText,
+				Flags: []cli.Flag{
+					&cli.StringFlag{
+						Name:     "vault-password-file",
+						Usage:    "vault password file `VAULT_PASSWORD_FILE`",
+						Required: false,
+					},
+					&cli.StringFlag{
+						Name:     "encrypted-env-file",
+						Aliases:  []string{"e"},
+						Usage:    "Path to encrypted Kube Secret file",
+						Required: true,
+					},
+					&cli.StringFlag{
+						Name:    "accessor",
+						Aliases: []string{"a"},
+						Usage:   "Accessor key to pull data out of Data block.",
+						Value:   ".env",
+					},
+					&cli.StringFlag{
+						Name:     "namespace",
+						Aliases:  []string{"n"},
+						Usage:    "Kube Namespace list Pods from",
+						Required: true,
+					},
+					&cli.StringFlag{
+						Name:     "dotenv",
+						Usage:    "Path to .env file on Pod",
+						Required: false,
+						Value:    "$PWD/.env",
+					},
+				},
+				Action: cmd.ViewAnsibleEnvDiff,
 			},
 		},
 	}
